requirements.txt
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
transformers[serving] @ git+https://github.com/huggingface/transformers.git@main
datasets>=2.21.0
trl>=0.10.1
accelerate>=0.34.0
peft>=0.12.0
scikit-learn>=1.5.1
pandas>=2.2.2
pyarrow>=17.0.0
evaluate>=0.4.3
rouge-score>=0.1.2
nltk>=3.9.1
matplotlib>=3.9.2
sentencepiece>=0.2.0
torchvision>=0.19.0
pillow>=10.4.0
jupyter>=1.1.1
ipykernel>=6.29.5
